move nettest to common/testing
